+++ src/main/java/fr/maxlego08/essentials/vault/VaultModule.java
@@ -242,8 +242,8 @@ public class VaultModule extends ZModule implements VaultManager {
     @Override
     public List<String> getVaultAsTabCompletion(Player player) {
         List<String> strings = IntStream.range(1, this.maxVaults).filter(vaultID -> hasPermission(player.getUniqueId(), vaultID)).mapToObj(String::valueOf).collect(Collectors.toList());
-        if (hasPermission(player, Permission.VAULT_ADD_SLOT)) strings.add("add");
-        if (hasPermission(player, Permission.VAULT_SET_SLOT)) strings.add("set");
+        if (hasPermission(player, Permission.ESSENTIALS_VAULT_ADD_SLOT)) strings.add("add");
+        if (hasPermission(player, Permission.ESSENTIALS_VAULT_SET_SLOT)) strings.add("set");
         return strings;
     }